Move files to match standard golang repo layout
